updated circular dependency fix
