+++ src/lib/finalizer.test.ts
@@ -6,7 +6,8 @@ import { addFinalizer, removeFinalizer } from "./finalizer";
 import { KubernetesObject, K8s, GenericClass, RegisterKind } from "kubernetes-fluent-client";
 import { K8sInit } from "kubernetes-fluent-client/dist/fluent/types";
 import { AdmissionRequest } from "./types";
-import { PeprMutateRequest, Operation } from "./mutate-types";
+import { Operation } from "./mutate-types";
+import { PeprMutateRequest } from "./mutate-request";
 import { Binding } from "./types";
 
 jest.mock("kubernetes-fluent-client");
