+++ src/lib/finalizer.ts
@@ -5,7 +5,7 @@ import { K8s, KubernetesObject, RegisterKind } from "kubernetes-fluent-client";
 import Log from "./logger";
 import { Binding } from "./types";
 import { Operation, DeepPartial } from "./mutate-types";
-import { PeprMutateRequest } from "./mutate-types";
+import { PeprMutateRequest } from "./mutate-request";
 
 export function addFinalizer<K extends KubernetesObject>(request: PeprMutateRequest<K>) {
   // if a DELETE is being processed, don't add a finalizer
