+++ src/lib/mutate-request.test.ts
@@ -4,7 +4,8 @@
 import { beforeEach, describe, expect, it } from "@jest/globals";
 import { KubernetesObject } from "kubernetes-fluent-client";
 import { AdmissionRequest } from "./types";
-import { PeprMutateRequest, Operation } from "./mutate-types";
+import { Operation } from "./mutate-types";
+import { PeprMutateRequest } from "./mutate-request";
 
 describe("PeprMutateRequest", () => {
   let mockRequest: AdmissionRequest<KubernetesObject>;
